% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_IBOSS.R
\name{GenIBOSS}
\alias{GenIBOSS}
\title{GenIBOSS}
\usage{
GenIBOSS(
  X = NULL,
  y = NULL,
  csv = NULL,
  nSample = -1,
  k = -1,
  family,
  intercept = FALSE,
  header = FALSE
)
}
\arguments{
\item{X}{Numeric matrix of predictors.}

\item{y}{Numeric response vector.}

\item{csv}{Path to CSV file (alternative to X, y).
Last column is assumed to be response.}

\item{nSample}{Number of rows to be subsampled for
initial estimate of parameters}

\item{k}{Integer; Number of points to select.}

\item{family}{Family of distribution}

\item{intercept}{Logical; whether first column is intercept.}

\item{header}{Logical; whether the csv files contains a header row.}
}
\value{
A list with:
\itemize{
\item final_model: result after performing glm on subsampled data.
}
}
\description{
Runs the IBOSS algorithm on given dataset.
}
