% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBOSS.R
\name{IBOSS}
\alias{IBOSS}
\title{IBOSS}
\usage{
IBOSS(X = NULL, y = NULL, csv = NULL, k, intercept = FALSE, header = FALSE)
}
\arguments{
\item{X}{Numeric matrix of predictors.}

\item{y}{Numeric response vector.}

\item{csv}{Path to CSV file (alternative to X, y).
Last column is assumed to be response.}

\item{k}{Integer; Number of points to select.}

\item{intercept}{Logical; whether first column is intercept.}

\item{header}{Logical; whether the csv files contains a header row.}
}
\value{
A list with:
\itemize{
\item X_selected: Numeric matrix of subset data
\item y_selected: Numeric vector of subset response
}
}
\description{
Runs the IBOSS algorithm on given dataset.
}
\examples{

set.seed (42)
X <- matrix(rnorm(200), ncol = 4)
y <- rnorm(50)
res <- IBOSS(X = X, y = y, k = 20)
str(res)



}
