% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRHT.R
\name{SRHT}
\alias{SRHT}
\title{Subsampled Randomized Hadamard Transform (SRHT)}
\usage{
SRHT(X = NULL, y = NULL, csv = NULL, k, intercept = FALSE, header = FALSE)
}
\arguments{
\item{X}{A numeric matrix of predictors. Required if \code{csv} is NULL.
Must have the same number of rows as \code{y}.}

\item{y}{A numeric vector of responses. Required if \code{csv} is NULL.}

\item{csv}{A character string specifying the path to a CSV file. If provided,
\code{X} and \code{y} must be NULL. The function assumes the last column
of the CSV is the response variable \code{y} and all preceding columns
are predictors \code{X}.}

\item{k}{An integer specifying the target subsample size.}

\item{intercept}{Boolean variable. If \code{TRUE}, the function checks if the first
column of \code{X} is all ones. If not, a column of ones is prepended
to standardizes the intercept term. Defaults to \code{FALSE}.}

\item{header}{Boolean variable. Indicates if the CSV file contains a header row.
Passed directly to \code{data.table::fread}. Defaults to \code{FALSE}.}
}
\value{
The result list returned by the underlying C++ \code{SRHT_cpp} function.
}
\description{
This function acts as a wrapper to prepare data and perform SRHT-based
regression. It handles input from either in-memory matrices
or CSV files, checks for an intercept, ensures correct data types, and
passes the data to the optimized C++ implementation.
}
